[#2257] Inject Environment Variable for Github Pages Deployments in Vite Build  (#16)
Inject Environment Variable for Github Pages Deployments in Vite Build

Under the Vite build for Github Pages Deployments, the base directory
of the deployment needs to be specified in a configuration file during
build time. We can make this dynamic since the base directory is the
name of the deployment repository.

Let's modify the code to add the VITE_BASE_DIR environment variable for
Github Pages Deployments.